pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "survey-system"
version = "0.1.0"
description = "AI-powered no-code survey system (Phase 1 POC)"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "httpx>=0.27",
    "sentence-transformers>=3.0",
    "faiss-cpu>=1.8",
    "numpy>=1.24",
    "streamlit>=1.34",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]
voice = [
    "faster-whisper>=1.0.3",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
